Add place order and confirmation

+++ apps/cart/views.py
@@ -5,22 +5,21 @@ from django.views.decorators.http import require_POST
 from apps.cart.models import Cart, CartItem
 from apps.products.models import Product
 
-# Create your views here.
-
 
 def cart_details(request):
     cart_id = request.session.get("cart_id")
-    cart = None
 
     if cart_id:
-        cart = get_object_or_404(Cart, id=cart_id)
+        cart = Cart.objects.filter(
+            id=cart_id
+        ).first()  # Usamos .filter().first() para evitar errores si no existe
+    else:
+        cart = None  # Si no hay cart_id en la sesión, el carrito es None
 
     return render(
         request,
         "cart/details.html",
-        {
-            "cart": cart,
-        },
+        {"cart": cart},
     )
 
 
@@ -28,15 +27,11 @@ def cart_details(request):
 def cart_add(request, product_id):
     cart_id = request.session.get("cart_id")
 
-    if not cart_id:
+    if not cart_id or not Cart.objects.filter(id=cart_id).exists():
         cart = Cart.objects.create()
         request.session["cart_id"] = cart.id
     else:
-        try:
-            cart = Cart.objects.get(id=cart_id)
-        except Cart.DoesNotExist:
-            cart = Cart.objects.create()
-            request.session["cart_id"] = cart.id
+        cart = Cart.objects.get(id=cart_id)
 
     product = get_object_or_404(Product, id=product_id)
     cart_item, created = CartItem.objects.get_or_create(cart=cart, product=product)
@@ -57,7 +52,11 @@ def cart_add(request, product_id):
 @require_POST
 def cart_remove_item(request, product_id):
     cart_id = request.session.get("cart_id")
-    cart = get_object_or_404(Cart, id=cart_id)
+
+    if not cart_id or not Cart.objects.filter(id=cart_id).exists():
+        return redirect("cart:cart_details")  # Evita error si no hay carrito
+
+    cart = Cart.objects.get(id=cart_id)
     item = get_object_or_404(CartItem, cart=cart, product_id=product_id)
     item.delete()
